messaging_app/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-nested-routers==0.95.0
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
PyJWT==2.10.1
sqlparse==0.5.3
uuid==1.30
